Done RegisterDesktop logic still to do proper styling
